+++ packages/angular/src/lib/Directives/DataTypes.ts
@@ -26,6 +26,12 @@ type WebtrekkAdvancedPropRequestQueue = {
     size?: NumberOrString;
 }
 
+type WebtrekkAdvancedPropRequestLimit = {
+    activated?: boolean;
+    amount?: NumberOrString;
+    duration?: NumberOrString;
+}
+
 type WebtrekkAdvancedPropUserIdentification = {
     enableOptOut?: boolean;
     enableAnonymousFunction?: boolean;
@@ -59,6 +65,7 @@ export interface WebtrekkAdvancedProps {
     useHashForDefaultPageName?: boolean;
     useParamsForDefaultPageName?: string[];
     requestQueue?: WebtrekkAdvancedPropRequestQueue;
+    requestLimit?: WebtrekkAdvancedPropRequestLimit;
     userIdentification?: WebtrekkAdvancedPropUserIdentification;
 }