+++ packages/vue/types/types.d.ts
@@ -29,6 +29,12 @@ type WebtrekkAdvancedPropRequestQueue = {
     size?: NumberOrString;
 }
 
+type WebtrekkAdvancedPropRequestLimit = {
+    activated?: boolean;
+    amount?: NumberOrString;
+    duration?: NumberOrString;
+}
+
 type WebtrekkAdvancedPropUserIdentification = {
     enableOptOut?: boolean;
     enableAnonymousFunction?: boolean;
@@ -56,6 +62,7 @@ interface WebtrekkAdvancedProps {
     useHashForDefaultPageName?: boolean;
     useParamsForDefaultPageName?: string[];
     requestQueue?: WebtrekkAdvancedPropRequestQueue;
+    requestLimit?: WebtrekkAdvancedPropRequestLimit;
     userIdentification?: WebtrekkAdvancedPropUserIdentification;
 }